PFA.3/requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
astroid==2.5.6
colorama==0.4.4
Django==3.0.5
django-graphql-jwt==0.3.0
djongo==1.3.3
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphviz==0.16
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
promise==2.3
PyJWT==2.1.0
pylint==2.8.2
pymongo==3.11.4
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
